Reorganised server code
